Chore: Fix UI colour component state management
Previous version required a hack that would not work in production
due to strict mode, where a ref was used to prevent a useEffect from
setting state on mount (only on update).
Now this has been fixed so the intended behaviour is identical both in
devlopment with strict mode and in production without.

Updated README.